scripts/requirements.txt
# scripts/bootstrap.py & compiler 依赖 (Python 3.11+)
# AI-generated: bootstrap 使用 Docker SDK、YAML 解析与完整性校验
PyYAML>=6.0
Jinja2>=3.1
docker>=7.0.0
